+++ CHANGELOG.md
@@ -21,6 +21,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Fixed code generation error when changing views in MS SQL
 - Fixed bug with ignoring column order in PostgreSQL and MS SQL constraints.
 - Fixed a bug with the settings when saving objects to a project.
+- Fixed false differences when ignoring table column order.
 
 ### Removed